+++ plugins/flux/src/components/panel/PanelList.tsx
@@ -49,7 +49,7 @@ const PanelList: React.FunctionComponent<IPanelListProps> = ({
         const json = await response.json();
 
         if (response.status >= 200 && response.status < 300) {
-          return resource.rows(json);
+          return resource.rows(json, undefined);
         } else {
           if (json.error) {
             throw new Error(json.error);